Cleaned up a few things. Got Tailwind working in Storybook! Decided not to prematurely organize components; nuked the App directory under src.
