+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/shawnwilson/dev/mcc-react/src/App.js":"1","/Users/shawnwilson/dev/mcc-react/src/App/Avatar.js":"2","/Users/shawnwilson/dev/mcc-react/src/stories/Page.stories.js":"3","/Users/shawnwilson/dev/mcc-react/src/stories/Button.js":"4","/Users/shawnwilson/dev/mcc-react/src/stories/Avatar.stories.js":"5","/Users/shawnwilson/dev/mcc-react/src/stories/Header.stories.js":"6","/Users/shawnwilson/dev/mcc-react/src/stories/Header.js":"7","/Users/shawnwilson/dev/mcc-react/src/index.js":"8","/Users/shawnwilson/dev/mcc-react/src/reportWebVitals.js":"9"},{"size":269,"mtime":1611019024386,"results":"10","hashOfConfig":"11"},{"size":546,"mtime":1611027699266,"results":"12","hashOfConfig":"11"},{"size":424,"mtime":1611001423218,"results":"13","hashOfConfig":"11"},{"size":1101,"mtime":1611001423216,"results":"14","hashOfConfig":"11"},{"size":414,"mtime":1611019186858,"results":"15","hashOfConfig":"11"},{"size":325,"mtime":1611001423217,"results":"16","hashOfConfig":"11"},{"size":1463,"mtime":1611001423217,"results":"17","hashOfConfig":"11"},{"size":506,"mtime":1611027763565,"results":"18","hashOfConfig":"11"},{"size":362,"mtime":1610997725907,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"12xnoq9",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"30","messages":"31","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/shawnwilson/dev/mcc-react/src/App.js",[],"/Users/shawnwilson/dev/mcc-react/src/App/Avatar.js",[],"/Users/shawnwilson/dev/mcc-react/src/stories/Page.stories.js",["38"],"/Users/shawnwilson/dev/mcc-react/src/stories/Button.js",[],"/Users/shawnwilson/dev/mcc-react/src/stories/Avatar.stories.js",["39"],"/Users/shawnwilson/dev/mcc-react/src/stories/Header.stories.js",["40"],"/Users/shawnwilson/dev/mcc-react/src/stories/Header.js",[],"/Users/shawnwilson/dev/mcc-react/src/index.js",[],"/Users/shawnwilson/dev/mcc-react/src/reportWebVitals.js",[],{"ruleId":"41","severity":1,"message":"42","line":6,"column":1,"nodeType":"43","endLine":9,"endColumn":3},{"ruleId":"41","severity":1,"message":"42","line":5,"column":1,"nodeType":"43","endLine":8,"endColumn":3},{"ruleId":"41","severity":1,"message":"42","line":5,"column":1,"nodeType":"43","endLine":8,"endColumn":3},"import/no-anonymous-default-export","Assign object to a variable before exporting as module default","ExportDefaultDeclaration"]
+[{"/Users/shawnwilson/dev/mcc-react/src/App.js":"1","/Users/shawnwilson/dev/mcc-react/src/stories/Page.stories.js":"2","/Users/shawnwilson/dev/mcc-react/src/stories/Button.js":"3","/Users/shawnwilson/dev/mcc-react/src/stories/Avatar.stories.js":"4","/Users/shawnwilson/dev/mcc-react/src/stories/Header.stories.js":"5","/Users/shawnwilson/dev/mcc-react/src/stories/Header.js":"6","/Users/shawnwilson/dev/mcc-react/src/index.js":"7","/Users/shawnwilson/dev/mcc-react/src/reportWebVitals.js":"8","/Users/shawnwilson/dev/mcc-react/src/stories/Page.js":"9","/Users/shawnwilson/dev/mcc-react/src/Avatar.js":"10"},{"size":266,"mtime":1611033274145,"results":"11","hashOfConfig":"12"},{"size":424,"mtime":1611001423218,"results":"13","hashOfConfig":"12"},{"size":1101,"mtime":1611001423216,"results":"14","hashOfConfig":"12"},{"size":495,"mtime":1611033248790,"results":"15","hashOfConfig":"12"},{"size":325,"mtime":1611001423217,"results":"16","hashOfConfig":"12"},{"size":1463,"mtime":1611001423217,"results":"17","hashOfConfig":"12"},{"size":506,"mtime":1611027763565,"results":"18","hashOfConfig":"12"},{"size":362,"mtime":1610997725907,"results":"19","hashOfConfig":"12"},{"size":2521,"mtime":1611001423218,"results":"20","hashOfConfig":"12"},{"size":682,"mtime":1611033259355,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"12xnoq9",{"filePath":"24","messages":"25","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"30","messages":"31","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/shawnwilson/dev/mcc-react/src/App.js",[],"/Users/shawnwilson/dev/mcc-react/src/stories/Page.stories.js",["42"],"/Users/shawnwilson/dev/mcc-react/src/stories/Button.js",[],"/Users/shawnwilson/dev/mcc-react/src/stories/Avatar.stories.js",["43"],"/Users/shawnwilson/dev/mcc-react/src/stories/Header.stories.js",["44"],"/Users/shawnwilson/dev/mcc-react/src/stories/Header.js",[],"/Users/shawnwilson/dev/mcc-react/src/index.js",[],"/Users/shawnwilson/dev/mcc-react/src/reportWebVitals.js",[],"/Users/shawnwilson/dev/mcc-react/src/stories/Page.js",[],"/Users/shawnwilson/dev/mcc-react/src/Avatar.js",[],{"ruleId":"45","severity":1,"message":"46","line":6,"column":1,"nodeType":"47","endLine":9,"endColumn":3},{"ruleId":"45","severity":1,"message":"46","line":5,"column":1,"nodeType":"47","endLine":8,"endColumn":3},{"ruleId":"45","severity":1,"message":"46","line":5,"column":1,"nodeType":"47","endLine":8,"endColumn":3},"import/no-anonymous-default-export","Assign object to a variable before exporting as module default","ExportDefaultDeclaration"]
