+++ .storybook/preview.js
@@ -1,3 +1,5 @@
+// Import our Tailwind styles
+import '../src/assets/main.css';
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
