+++ package.json
@@ -19,8 +19,8 @@
     "build": "npm run watch:css && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public",
+    "storybook": "npm run watch:css && start-storybook -p 6006 -s public",
+    "build-storybook": "npm run watch:css && build-storybook -s public",
     "watch:css": "postcss src/assets/tailwind.css -o src/assets/main.css"
   },
   "eslintConfig": {
